Exit with a non-zero exit code if nothing is logged to STDERR.
A proper exit code is needed for monitoring of the success or failure of
the backup attempt by services like probyapp.com.

It is possible to get mongodump to not print to STDERR by making sure
that all options are configured correctly (for example, disabling
replication finding if not using replica sets).