Exit if no frequency is set
This was causing problems when someone tried to disable our backups.

The expected behaviour (no backups) didn't happen :(

Instead the script just backs up without FILE variable being set which
causes the mongodump to run into the the CWD which is worse.

Exit with non zero return code seems best as when in a cron we get a
message to say that it failed. It's probably best to be bugged when
backsup are not running as it should be an execptional config state.